Convert getVisibilityDescription and getVisibilityIcon to Visibility? extension functions
