+++ app/src/main/java/app/pachli/adapter/StatusDetailedViewHolder.kt
@@ -1,6 +1,5 @@
 package app.pachli.adapter
 
-import android.graphics.drawable.Drawable
 import android.os.Build
 import android.text.SpannableStringBuilder
 import android.text.Spanned
@@ -8,11 +7,11 @@ import android.text.method.LinkMovementMethod
 import android.text.style.DynamicDrawableSpan
 import android.text.style.ImageSpan
 import android.view.View
-import androidx.appcompat.content.res.AppCompatResources
 import androidx.recyclerview.widget.RecyclerView
 import app.pachli.R
 import app.pachli.databinding.ItemStatusDetailedBinding
-import app.pachli.entity.Status
+import app.pachli.entity.description
+import app.pachli.entity.icon
 import app.pachli.interfaces.StatusActionListener
 import app.pachli.util.CardViewMode
 import app.pachli.util.NoUnderlineURLSpan
@@ -34,8 +33,8 @@ class StatusDetailedViewHolder(
     ) {
         val (_, _, _, _, _, _, _, createdAt, editedAt, _, _, _, _, _, _, _, _, _, visibility, _, _, _, app) = statusViewData.actionable
         val context = metaInfo.context
-        val visibilityIcon = getVisibilityIcon(visibility)
-        val visibilityString = getVisibilityDescription(context, visibility)
+        val visibilityIcon = visibility.icon(metaInfo)
+        val visibilityString = visibility.description(context)
         val sb = SpannableStringBuilder(visibilityString)
         if (visibilityIcon != null) {
             val visibilityIconSpan = ImageSpan(
@@ -146,34 +145,6 @@ class StatusDetailedViewHolder(
                 hideQuantitativeStats()
             }
         }
-    }
-
-    private fun getVisibilityIcon(visibility: Status.Visibility?): Drawable? {
-        if (visibility == null) {
-            return null
-        }
-        val visibilityIcon: Int = when (visibility) {
-            Status.Visibility.PUBLIC -> R.drawable.ic_public_24dp
-            Status.Visibility.UNLISTED -> R.drawable.ic_lock_open_24dp
-            Status.Visibility.PRIVATE -> R.drawable.ic_lock_outline_24dp
-            Status.Visibility.DIRECT -> R.drawable.ic_email_24dp
-            else -> {
-                return null
-            }
-        }
-        val visibilityDrawable = AppCompatResources.getDrawable(
-            metaInfo.context,
-            visibilityIcon,
-        ) ?: return null
-        val size = metaInfo.textSize.toInt()
-        visibilityDrawable.setBounds(
-            0,
-            0,
-            size,
-            size,
-        )
-        visibilityDrawable.setTint(metaInfo.currentTextColor)
-        return visibilityDrawable
     }
 
     private fun hideQuantitativeStats() {
